+++ Source/OCMockito/MKTOngoingStubbing.m
@@ -120,11 +120,11 @@
     return self;
 }
 
-- (MKTOngoingStubbing *)willReturnStruct:(void *)value
+- (MKTOngoingStubbing *)willReturnStruct:(const void *)value objCType:(const char *)type
 {
-  NSValue *answer = [NSValue valueWithBytes:value objCType:@encode(typeof(value))];
-  [_invocationContainer addAnswer:answer];
-  return self;
+    NSValue *answer = [NSValue valueWithBytes:value objCType:type];
+    [_invocationContainer addAnswer:answer];
+    return self;
 }